+++ src/anyio/__init__.py
@@ -7,11 +7,12 @@ __all__ = ('maybe_async', 'maybe_async_cm', 'run', 'sleep', 'current_time', 'get
            'create_connected_udp_socket', 'getaddrinfo', 'getnameinfo', 'wait_socket_readable',
            'wait_socket_writable', 'create_memory_object_stream', 'run_process', 'open_process',
            'current_default_worker_process_limiter', 'run_sync_in_process', 'create_lock',
-           'CapacityLimiterStatistics', 'ConditionStatistics', 'EventStatistics', 'LockStatistics',
+           'CapacityLimiter', 'CapacityLimiterStatistics', 'Condition', 'ConditionStatistics',
+           'Event', 'EventStatistics', 'Lock', 'LockStatistics', 'Semaphore',
            'SemaphoreStatistics', 'create_condition', 'create_event', 'create_semaphore',
            'create_capacity_limiter', 'open_cancel_scope', 'fail_after', 'move_on_after',
-           'current_effective_deadline', 'TASK_STATUS_IGNORED', 'create_task_group', 'TaskInfo',
-           'get_current_task', 'get_running_tasks', 'wait_all_tasks_blocked',
+           'current_effective_deadline', 'TASK_STATUS_IGNORED', 'CancelScope', 'create_task_group',
+           'TaskInfo', 'get_current_task', 'get_running_tasks', 'wait_all_tasks_blocked',
            'run_sync_in_worker_thread', 'run_async_from_thread', 'run_sync_from_thread',
            'current_default_worker_thread_limiter', 'create_blocking_portal',
            'start_blocking_portal', 'typed_attribute', 'TypedAttributeSet',
@@ -33,17 +34,17 @@ from ._core._streams import create_memory_object_stream
 from ._core._subprocesses import (
     current_default_worker_process_limiter, open_process, run_process, run_sync_in_process)
 from ._core._synchronization import (
-    ConditionStatistics, LockStatistics, SemaphoreStatistics, create_capacity_limiter,
+    CapacityLimiter, CapacityLimiterStatistics, Condition, ConditionStatistics, Event,
+    EventStatistics, Lock, LockStatistics, Semaphore, SemaphoreStatistics, create_capacity_limiter,
     create_condition, create_event, create_lock, create_semaphore)
 from ._core._tasks import (
-    TASK_STATUS_IGNORED, create_task_group, current_effective_deadline, fail_after, move_on_after,
-    open_cancel_scope)
+    TASK_STATUS_IGNORED, CancelScope, create_task_group, current_effective_deadline, fail_after,
+    move_on_after, open_cancel_scope)
 from ._core._testing import TaskInfo, get_current_task, get_running_tasks, wait_all_tasks_blocked
 from ._core._threads import (
     create_blocking_portal, current_default_worker_thread_limiter, run_async_from_thread,
     run_sync_from_thread, run_sync_in_worker_thread, start_blocking_portal)
 from ._core._typedattr import TypedAttributeProvider, TypedAttributeSet, typed_attribute
-from .abc._synchronization import CapacityLimiterStatistics, EventStatistics
 
 # Re-export imports so they look like they live directly in this package
 for key, value in list(locals().items()):