+++ src/anyio/_backends/_asyncio.py
@@ -29,7 +29,10 @@ from .._core._exceptions import (
 from .._core._exceptions import ExceptionGroup as BaseExceptionGroup
 from .._core._exceptions import WouldBlock
 from .._core._sockets import GetAddrInfoReturnType, convert_ipv6_sockaddr
+from .._core._synchronization import CapacityLimiter as BaseCapacityLimiter
+from .._core._synchronization import Event as BaseEvent
 from .._core._synchronization import ResourceGuard
+from .._core._tasks import CancelScope as BaseCancelScope
 from ..abc import IPSockAddrType, UDPPacketType
 from ..lowlevel import RunVar
 
@@ -220,9 +223,9 @@ sleep = asyncio.sleep
 CancelledError = asyncio.CancelledError
 
 
-class CancelScope(abc.CancelScope, DeprecatedAsyncContextManager):
-    __slots__ = ('_deadline', '_shield', '_parent_scope', '_cancel_called', '_active',
-                 '_timeout_handle', '_tasks', '_host_task', '_timeout_expired')
+class CancelScope(BaseCancelScope, DeprecatedAsyncContextManager):
+    def __new__(cls, *, deadline: float = math.inf, shield: bool = False):
+        return object.__new__(cls)
 
     def __init__(self, deadline: float = math.inf, shield: bool = False):
         self._deadline = deadline
@@ -1523,7 +1526,10 @@ async def wait_socket_writable(sock: socket.SocketType) -> None:
 # Synchronization
 #
 
-class Event(abc.Event):
+class Event(BaseEvent):
+    def __new__(cls):
+        return object.__new__(cls)
+
     def __init__(self):
         self._event = asyncio.Event()
 
@@ -1542,8 +1548,11 @@ class Event(abc.Event):
         return EventStatistics(len(self._event._waiters))
 
 
-class CapacityLimiter(abc.CapacityLimiter):
+class CapacityLimiter(BaseCapacityLimiter):
     _total_tokens: float = 0
+
+    def __new__(cls, total_tokens: float):
+        return object.__new__(cls)
 
     def __init__(self, total_tokens: float):
         self._borrowers: Set[Any] = set()