+++ src/anyio/_backends/_trio.py
@@ -21,7 +21,10 @@ from .._core._exceptions import (
     BrokenResourceError, BusyResourceError, ClosedResourceError, EndOfStream)
 from .._core._exceptions import ExceptionGroup as BaseExceptionGroup
 from .._core._sockets import convert_ipv6_sockaddr
+from .._core._synchronization import CapacityLimiter as BaseCapacityLimiter
+from .._core._synchronization import Event as BaseEvent
 from .._core._synchronization import ResourceGuard
+from .._core._tasks import CancelScope as BaseCancelScope
 from ..abc import IPSockAddrType, UDPPacketType
 
 try:
@@ -56,7 +59,10 @@ sleep = trio.sleep
 # Timeouts and cancellation
 #
 
-class CancelScope(abc.CancelScope):
+class CancelScope(BaseCancelScope):
+    def __new__(cls, original: Optional[trio.CancelScope] = None, **kwargs):
+        return object.__new__(cls)
+
     def __init__(self, original: Optional[trio.CancelScope] = None, **kwargs):
         self.__original = original or trio.CancelScope(**kwargs)
 
@@ -113,7 +119,7 @@ class TaskGroup(abc.TaskGroup):
     async def __aenter__(self):
         self._active = True
         self._nursery = await self._nursery_manager.__aenter__()
-        self.cancel_scope = self._nursery.cancel_scope
+        self.cancel_scope = CancelScope(self._nursery.cancel_scope)
         return self
 
     async def __aexit__(self, exc_type: Optional[Type[BaseException]],
@@ -569,7 +575,10 @@ async def wait_socket_writable(sock):
 # Synchronization
 #
 
-class Event(abc.Event):
+class Event(BaseEvent):
+    def __new__(cls):
+        return object.__new__(cls)
+
     def __init__(self):
         self.__original = trio.Event()
 
@@ -587,9 +596,12 @@ class Event(abc.Event):
         return DeprecatedAwaitable(self.set)
 
 
-class CapacityLimiter(abc.CapacityLimiter):
-    def __init__(self, *args, original: Optional[trio.CapacityLimiter] = None, **kwargs):
-        self.__original = original or trio.CapacityLimiter(*args, **kwargs)
+class CapacityLimiter(BaseCapacityLimiter):
+    def __new__(cls, *args, **kwargs):
+        return object.__new__(cls)
+
+    def __init__(self, *args, original: Optional[trio.CapacityLimiter] = None):
+        self.__original = original or trio.CapacityLimiter(*args)
 
     async def __aenter__(self):
         return await self.__original.__aenter__()